backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voiceflow-backend"
version = "0.1.0"
description = "VoiceFlow CMS Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.114.0",
    "uvicorn[standard]==0.30.5",
    "orjson==3.10.7",
    "python-multipart==0.0.9",
    "pydantic-settings==2.4.0",
    "SQLAlchemy==2.0.32",
    "asyncpg==0.29.0",
    "alembic==1.13.2",
    "psycopg[binary]==3.1.19",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.2.0",
    "redis==5.0.7",
    "celery==5.4.0",
    "cloudinary==1.41.0",
    "openai-whisper==20231117",
    "librosa==0.10.2.post1",
    "soundfile==0.12.1",
    "numpy==1.26.4",
    "pydub==0.25.1",
    "speechbrain==0.5.16",
    "torch==2.3.1+cpu",
    "torchaudio==2.3.1+cpu",
    "torchvision==0.18.1+cpu",
    "gTTS==2.5.1",
    "spacy==3.7.5",
    "rapidfuzz==3.9.6",
    "lingua-language-detector==1.3.2",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
]


